+++ udata_front/theme/gouvfr/datagouv-components/CHANGELOG.md
@@ -2,7 +2,18 @@
 
 ## Current (in progress)
 
-- Nothing yet
+- Use resource downloads for download icon on dataset card [#484](https://github.com/datagouv/udata-front/pull/484)
+- Reduce organization type icon [#483](https://github.com/datagouv/udata-front/pull/483)
+- add Tabs component [#491](https://github.com/datagouv/udata-front/pull/491)
+- show permalink for community resources [#486](https://github.com/datagouv/udata-front/pull/486)
+- fix dataset card tooltip bug [#493](https://github.com/datagouv/udata-front/pull/493)
+
+## 1.2.0 (2024-08-21)
+
+- Add new resources_dowloads metric to Dataset type [#474](https://github.com/datagouv/udata-front/pull/474)
+- Improve organization type logic [#475](https://github.com/datagouv/udata-front/pull/475)
+- Update reuse card [#449](https://github.com/datagouv/udata-front/pull/449)
+- Add dataservice card [#450](https://github.com/datagouv/udata-front/pull/450)
 
 ## 1.1.2 (2024-08-01)